fix plotting services
