+++ ooiservices/tests/test_uframe_data.py
@@ -27,6 +27,18 @@ class UframeDataTestCase(unittest.TestCase):
         db.session.remove()
         db.drop_all()
         self.app_context.pop()
+
+    def test_simple_fail_data_access_no_info(self):
+        response = self.client.get('/uframe/get_data', content_type='application/json')
+        self.assertTrue(response.status_code == 200)
+
+    def test_simple_fail_data_access_(self):
+        instrument = "CP02PMUO-WFP01-04-FLORTK000"
+        stream = "flort_kn_stc_imodem_instrument_broken_invalid"
+
+        response = self.client.get('/uframe/get_data', content_type='application/json')
+        self.assertTrue(response.status_code == 200)
+        
 ''' TODO: rewrite tests to reflect data from uframe
     def test_get_data_from_uframe_simple_ok(self):
         response = self.client.get('/uframe/get_data', content_type='application/json')
